[BE] refactor: 좋아요 증감 중복 요청 및 응원하기 버그 수정 (#701)
* feat: 쿠키 생성 API 추가

* feat: 특정 유저가 좋아요한 피드백들을 관리하는 저장소 구현

* feat: 좋아요 쿠키 기능 구현

- 이미 좋아요를 누른 사람은, 좋아요를 누를 수 없도록 쿠키를 사용해 구현함

* refactor: saveUser 메서드 삭제

- saveUser 메서드 대신, save 시점에 유저가 없다면 삽입하는 코드로 수정하였음

* test: 기존 테스트 코드 수정

- 테스트 코드에 쿠키를 삽입하였음

* test: 좋아요 중복 방지 테스트 코드 작성

* refactor: 좋아요 중복 오류 custom exception 도입

* refactor: 좋아요 취소 시 쿠키값을 통해 검증하도록 수정

- 쿠키가 존재하지 않거나, 쿠키값에 좋아요 기록이 없으면 좋아요 취소를 할 수 없도록 수정하였음

* refactor: 좋아요 취소 테스트 코드 작성 및 메서드 네이밍 수정

- unLike -> unlike 으로 수정하였음

* refactor: 응원하기 버그 값 수정

- int -> long 타입으로 변경
- 1,000이 넘는 값이 들어왔을 때, 1,000으로 치환해서 저장하도록 수정

* refactor: coderabbit 리뷰 반영

- 동시성 보장 위한 concurrentHashMap 도입
- cheeringCount nullable false 적용

* refactor: 타입 변경으로 인한 flyway script 작성

* refactor: 비즈니스 요구사항 변경으로 인한 응원하기 로직 수정

- 응원을 한 번에 100을 초과할 경우, 예외가 발생한다

* refactor: 쿠키 생성 로직 변경

* refactor: set-cookie responseEntity 변경

* refactor: visitorId null 처리 추가

* test: feedbackLikeServiceTest 코드 수정

- 잘못된 when/then 수정
- 사용하지 않는 import문 삭제

* refactor: flywayScript 버전 변경

+++ backend/src/test/java/feedzupzup/backend/organization/controller/OrganizationControllerE2ETest.java
@@ -38,7 +38,7 @@ class OrganizationControllerE2ETest extends E2EHelper {
                 .body("status", equalTo(200))
                 .body("message", equalTo("OK"))
                 .body("data.organizationName", equalTo(organization.getName().getValue()))
-                .body("data.totalCheeringCount", equalTo(organization.getCheeringCountValue()));
+                .body("data.totalCheeringCount", equalTo((int) organization.getCheeringCountValue()));
     }
 
     @Test
